Changed blastn to blastn-short and removed the word size to prevent blast results from having to have long strings of matches with no mismatch in the middle. Additionally, changed miRBaseMirnas.fa to miRBasePlantMirnas.fa so that non plant miRNAs are used in our annotation analysis. This helps with weird naming problems in some other organism miRNAs. Finally, added a check in housekeeping.py to ensure the user doesn't try to use the library input folder as the output folder to ensure their libraries do not get deleted.

+++ miRador/housekeeping.py
@@ -97,6 +97,10 @@ def housekeeping(genomeFilename, libFilenamesString, libFolder,
     # If the user has filled the outputFolder option, check to see if it
     # has results from an older run and then delete them
     if(outputFolder):
+        if(outputFolder == libFolder):
+            print("outputFolder and libFolder cannot be the same folder. "\
+                "Please rename outputFolder and run again")
+            sys.exit()
         if(os.path.isdir("%s/libs" % outputFolder)):
             shutil.rmtree("%s/libs" % outputFolder)
         if(os.path.isdir("%s/images" % outputFolder)):
